🔨 app: file rename for better consistency
