+++ app/src/App.tsx
@@ -8,10 +8,10 @@ import { HotKeys } from 'react-hotkeys'
 import { Provider } from 'react-redux'
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
 
-import { prepareModalState } from './components/CodeModal/functions'
+import { prepareModalState } from './components/CodeModal'
 import { Sidebar } from './components/Sidebar/Sidebar'
 import { Topbar } from './components/Topbar/Topbar'
-import { Dash } from './pages/Dash/DashContainer'
+import { Dash } from './pages/Dash/Dash'
 import { Login } from './pages/Login/Login'
 import { store } from './store/index.js'
 import { hotkeyHandlers, hotkeyMap } from './util/hotkeys'
